Package Update: Busybox: Allow L4T device builds to just dump files in initramfs, such as firmware or new init script

+++ packages/sysutils/busybox/package.mk
@@ -209,6 +209,19 @@ makeinstall_init() {
     touch $INSTALL/etc/fstab
     ln -sf /proc/self/mounts $INSTALL/etc/mtab
 
+  #Hack to swap out init, and add early firmware to initramfs for L4T builds
+  if [ ${PROJECT} == "L4T" ]; then
+    # Copy PROJECT related files to filesystem
+    if [ -d "${PROJECT_DIR}/${PROJECT}/initramfs" ]; then
+      cp -PR $PROJECT_DIR/$PROJECT/initramfs/* $INSTALL
+    fi
+
+    # Copy DEVICE related initramfs files to initramfs filesystem
+    if [ -n "$DEVICE" -a -d "$PROJECT_DIR/$PROJECT/devices/$DEVICE/initramfs" ]; then
+      cp -PR $PROJECT_DIR/$PROJECT/devices/$DEVICE/initramfs/* $INSTALL
+    fi
+  fi
+  
   if find_file_path initramfs/platform_init; then
     cp ${FOUND_PATH} $INSTALL
     sed -e "s/@BOOT_LABEL@/$DISTRO_BOOTLABEL/g" \
@@ -218,7 +231,9 @@ makeinstall_init() {
   fi
 
   cp $PKG_DIR/scripts/functions $INSTALL
-  cp $PKG_DIR/scripts/init $INSTALL
+  if [ ! -f $INSTALL/init ]; then
+    cp $PKG_DIR/scripts/init $INSTALL
+  fi
   sed -e "s/@DISTRONAME@/$DISTRONAME/g" \
       -e "s/@KERNEL_NAME@/$KERNEL_NAME/g" \
       -i $INSTALL/init
